Move second level browse page to own controller
