+++ app/controllers/top_level_browse_page_controller.rb
@@ -27,7 +27,7 @@ private
   end
 
   def second_level_browse_pages_partial(page)
-    render_partial('browse/_second_level_browse_pages',
+    render_partial('second_level_browse_page/_second_level_browse_pages',
       title: page.title,
       second_level_browse_pages: page.second_level_browse_pages,
       curated_order: page.second_level_pages_curated?,
