+++ app/views/second_level_browse_page/show.html.erb
@@ -24,6 +24,6 @@
       curated_order: @page.second_level_pages_curated? %>
   </div>
 
-  <%= render 'top_level_browse_pages',
+  <%= render 'browse/top_level_browse_pages',
     top_level_browse_pages: @page.top_level_browse_pages %>
 </div>
