+++ app/views/top_level_browse_page/show.html.erb
@@ -12,7 +12,7 @@
 
 <div class="browse-panes section" data-state="section">
   <div id="section" class="section-pane pane with-sort">
-    <%= render 'browse/second_level_browse_pages',
+    <%= render 'second_level_browse_page/second_level_browse_pages',
           title: @page.title,
           second_level_browse_pages: @page.second_level_browse_pages,
           curated_order: @page.second_level_pages_curated? %>
